Moved ChessPlot class to separate module
Renamed Game class to Interpreter

Made internal classes explicit with underscore

Added custom exceptions for interpreter errors
